Modify models and try to create a new buy page view controller because the seperate one is causing issues.
