+++ EngineYard_iOS/ViewControllers/BuyProductionViewController.swift
@@ -10,68 +10,16 @@ import UIKit
 
 class BuyProductionViewController: UIViewController {
 
-    weak var controller: TrainViewController?
     var productionPageViewModel: ProductionPageViewModel!
-
-    deinit {
-        guard let hasChildController = self.controller else {
-            return
-        }
-        hasChildController.removeFromParentViewController()
-    }
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        addTrainViewController()
     }
 
     override func didReceiveMemoryWarning() {
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
-
-    private func addTrainViewController() {
-        guard let hasViewModel = self.productionPageViewModel else {
-            assertionFailure("No view model defined")
-            return
-        }
-        guard let gameObj = hasViewModel.game else {
-            assertionFailure("No game object defined")
-            return
-        }
-
-        let sb: UIStoryboard = UIStoryboard(name: "Train", bundle: nil)
-        if let controller = sb.instantiateViewController(withIdentifier: "TrainViewController") as? TrainViewController
-        {
-            self.controller = controller
-            self.controller?.trainsViewModel = TrainListViewModel.init(game: gameObj)
-
-            addChildViewController(controller)
-            controller.view.translatesAutoresizingMaskIntoConstraints = false
-            view.addSubview(controller.view)
-
-            NSLayoutConstraint.activate([
-                controller.view.leadingAnchor.constraint(equalTo: self.view.leadingAnchor),
-                controller.view.trailingAnchor.constraint(equalTo: self.view.trailingAnchor),
-                controller.view.topAnchor.constraint(equalTo: self.view.topAnchor),
-                controller.view.bottomAnchor.constraint(equalTo: self.view.bottomAnchor)
-                ])
-
-            controller.didMove(toParentViewController: self)
-
-            controller.completionClosure = { (doneBtnPressed: Bool) in
-                // end turn, handle whether to move to next page
-                print ("doneBtn pressed")
-            }
-            controller.selectedTrainClosure = { (purchasedTrain: Locomotive?) in
-                print ("selectedTrainClosure pressed")
-            }
-        }
-    }
-
-
-
 
     // MARK: - Navigation
